separated base logic for oauth flow in order that can be used by dashboard
