+++ http_handlers.go
@@ -5,10 +5,10 @@ import (
 	"errors"
 	"fmt"
 	"github.com/TykTechnologies/tyk-identity-broker/constants"
+	"github.com/TykTechnologies/tyk-identity-broker/providers"
 	"net/http"
 
 	"github.com/Sirupsen/logrus"
-	"github.com/TykTechnologies/tyk-identity-broker/tap"
 	"github.com/gorilla/mux"
 )
 
@@ -62,7 +62,7 @@ func HandleAuth(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	thisIdentityProvider, err := tap.GetTapProfile(w, r, AuthConfigStore, IdentityKeyStore, thisId, TykAPIHandler)
+	thisIdentityProvider, err := providers.GetTapProfile(w, r, AuthConfigStore, IdentityKeyStore, thisId, TykAPIHandler)
 	if err != nil {
 		return
 	}
@@ -80,7 +80,7 @@ func HandleAuthCallback(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	thisIdentityProvider, err := tap.GetTapProfile(w, r, AuthConfigStore, thisId)
+	thisIdentityProvider, err := providers.GetTapProfile(w, r, AuthConfigStore, IdentityKeyStore, thisId, TykAPIHandler)
 	if err != nil {
 		HandleError(constants.HandlerLogTag, err.Message, err.Error, err.Code, w, r)
 		return
