+++ tap/profileActions.go
@@ -1,15 +1,8 @@
 package tap
 
 import (
-	"encoding/json"
 	"errors"
-	"github.com/Sirupsen/logrus"
-	"github.com/TykTechnologies/tyk-identity-broker/constants"
 	logger "github.com/TykTechnologies/tyk-identity-broker/log"
-	"github.com/TykTechnologies/tyk-identity-broker/providers"
-	identityHandlers "github.com/TykTechnologies/tyk-identity-broker/tap/identity-handlers"
-	"github.com/TykTechnologies/tyk-identity-broker/tyk-api"
-	"net/http"
 )
 
 var log = logger.Get()
@@ -117,79 +110,4 @@ func DeleteProfile(key string,AuthConfigStore AuthRegisterBackend, flush func(ba
 	return nil
 }
 
-func GetTapProfile(w http.ResponseWriter, r *http.Request, AuthConfigStore, identityKeyStore AuthRegisterBackend, id string,tykHandler tyk.TykAPI) (TAProvider, *HttpError) {
 
-	thisProfile := Profile{}
-	log.WithField("prefix", constants.HandlerLogTag).Debug("--> Looking up profile ID: ", id)
-	foundProfileErr := AuthConfigStore.GetKey(id, &thisProfile)
-
-	if foundProfileErr != nil {
-		errorMsg := "Profile " + id + " not found"
-		return nil, &HttpError{
-			Message: errorMsg,
-			Code:    404,
-			Error:   foundProfileErr,
-		}
-	}
-
-	thisIdentityProvider, providerErr := GetTAProvider(thisProfile,tykHandler,identityKeyStore)
-	if providerErr != nil {
-		return  nil, &HttpError{
-			Message: "Could not initialise provider",
-			Code:    400,
-			Error:   providerErr,
-		}
-	}
-
-	return thisIdentityProvider, nil
-}
-
-// return a provider based on the name of the provider type, add new providers here
-func GetTAProvider(conf Profile,handler tyk.TykAPI, identityKeyStore AuthRegisterBackend) (TAProvider, error) {
-
-	var thisProvider TAProvider
-
-	switch conf.ProviderName {
-	case constants.SocialProvider:
-		thisProvider = &providers.Social{}
-	case constants.ADProvider:
-		thisProvider = &providers.ADProvider{}
-	case constants.ProxyProvider:
-		thisProvider = &providers.ProxyProvider{}
-	default:
-		return nil, errors.New("invalid provider name")
-	}
-
-	thisIdentityHandler := getIdentityHandler(conf.ActionType, handler, identityKeyStore)
-	thisIdentityHandler.Init(conf)
-	err := thisProvider.Init(thisIdentityHandler, conf, hackProviderConf(conf.ProviderConfig))
-
-	return thisProvider, err
-}
-
-// Maps an identity handler from an Action type, register new Identity Handlers and methods here
-func getIdentityHandler(name Action,handler tyk.TykAPI, identityKeyStore AuthRegisterBackend) IdentityHandler {
-	var thisIdentityHandler IdentityHandler
-
-	switch name {
-	case GenerateOrLoginDeveloperProfile, GenerateOrLoginUserProfile, GenerateOAuthTokenForClient, GenerateTemporaryAuthToken:
-		thisIdentityHandler = &identityHandlers.TykIdentityHandler{
-			API:   &handler,
-			Store: identityKeyStore}
-	}
-
-	return thisIdentityHandler
-}
-
-// A hack to marshal a provider conf from map[string]interface{} into a type without type checking, ugly, but effective
-func hackProviderConf(conf interface{}) []byte {
-	thisConf, err := json.Marshal(conf)
-	if err != nil {
-		log.WithFields(logrus.Fields{
-			"prefix": constants.HandlerLogTag,
-			"error":  err,
-		}).Warning("Failure in JSON conversion")
-		return []byte{}
-	}
-	return thisConf
-}
